test-drm-prime-dumb-kms: add test program
This program allocates a dumb buffer with kms, exports and exports it
with prime. It then maps the buffer to user space and uses the CPU to
render some colors to it.
It is here to provide a reference to ensure a kms working driver with
prime before we add and accelerated graphics stuff and buffer sharing to
the mix.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>